Added editor GUI, made the scene serializable and added part of scripting system.
Added editor GUI, made the scene can be render in the window of GUI. And made the scene can be saved as a file (so far it is not binary file, because this is easy to debug). Last but not least, added the c# scripting system, it can basically link the c# binary file to invoke its code or make it invoke c++'s code.